vello_hybrid: Move `resource` module into custom file
